+++ packages/api/src/crud/auth/events/user-registration-completed.event.ts
