+++ packages/api/src/crud/auth/handlers/user-registration-started.handler.ts
