+++ packages/api/src/crud/courses/courses.controller.ts
@@ -2,8 +2,8 @@ import { Body, Controller, Get, Post, UseGuards } from '@nestjs/common';
 
 import { CourseBody, CreateCourseResponse, GetAllCoursesResponse } from '@coderscamp/shared/models/course';
 
+import { ForAdmin, RolesGuard } from '../auth/guards';
 import { JwtAuthGuard } from '../auth/jwt/jwt-auth.guard';
-import { ForAdmin, RolesGuard } from '../guards';
 import { CoursesService } from './courses.service';
 
 @ForAdmin()
