+++ packages/api/src/crud/courses/courses.module.ts
