+++ packages/api/src/module/read/course-progress/course-progress.read-module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { OnEvent } from '@nestjs/event-emitter';
 
+import { PrismaService } from '@/common/prisma/prisma.service';
 import { LearningMaterialsUrlWasGenerated } from '@/events/learning-materials-url-was-generated.domain-event';
 import { ApplicationEvent } from '@/module/application-command-events';
 import { TaskWasCompleted } from '@/module/events/task-was-completed.domain-event';
-import { PrismaService } from '@/prisma/prisma.service';
 import { SharedModule } from '@/write/shared/shared.module';
 
 import { CourseProgressRestController } from './course-progress.rest-controller';
