+++ packages/api/src/module/read/learning-materials/learning-materials.rest-controller.ts
@@ -2,10 +2,11 @@ import { Controller, Get, NotFoundException, UseGuards } from '@nestjs/common';
 
 import { GetLearningMaterialResponse } from '@coderscamp/shared/models/learning-material';
 
-import { JwtAuthGuard } from '@/auth/jwt/jwt-auth.guard';
-import { JwtUserId } from '@/auth/jwt/jwt-user-id.decorator';
-import { PrismaService } from '@/prisma/prisma.service';
-import { UserId } from '@/users/users.types';
+import { PrismaService } from '@/common/prisma/prisma.service';
+
+import { JwtAuthGuard } from '../../../crud/auth/jwt/jwt-auth.guard';
+import { JwtUserId } from '../../../crud/auth/jwt/jwt-user-id.decorator';
+import { UserId } from '../../../crud/users/users.types';
 
 @UseGuards(JwtAuthGuard)
 @Controller('learning-materials')
