+++ packages/api/src/module/write/learning-materials-url/infrastructure/users.adapter.ts
@@ -1,8 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 
-import { UsersService } from '@/users/users.service';
-import type { UserId } from '@/users/users.types';
-
+import { UsersService } from '../../../../crud/users/users.service';
+import type { UserId } from '../../../../crud/users/users.types';
 import { UsersPort } from '../application/users.port';
 
 @Injectable()
