+++ packages/api/src/module/write/learning-materials-url/presentation/rest/learning-materials-url.rest-controller.ts
@@ -1,11 +1,12 @@
 import { Controller, HttpCode, Post, UseGuards } from '@nestjs/common';
 import { CommandBus } from '@nestjs/cqrs';
 
-import { JwtAuthGuard } from '@/auth/jwt/jwt-auth.guard';
-import { JwtUserId } from '@/auth/jwt/jwt-user-id.decorator';
 import { GenerateLearningMaterialsUrlApplicationCommand } from '@/commands/generate-learning-materials-url.application-command';
-import { UserId } from '@/users/users.types';
 import { ApplicationCommandFactory } from '@/write/shared/application/application-command.factory';
+
+import { JwtAuthGuard } from '../../../../../crud/auth/jwt/jwt-auth.guard';
+import { JwtUserId } from '../../../../../crud/auth/jwt/jwt-user-id.decorator';
+import { UserId } from '../../../../../crud/users/users.types';
 
 @Controller('learning-materials')
 export class LearningMaterialsUrlRestController {
