+++ packages/api/src/module/write/shared/shared.module.ts
@@ -3,7 +3,7 @@ import { ModuleMetadata } from '@nestjs/common/interfaces/modules/module-metadat
 import { CqrsModule } from '@nestjs/cqrs';
 
 import { env } from '@/common/env';
-import { PrismaModule } from '@/prisma/prisma.module';
+import { PrismaModule } from '@/common/prisma/prisma.module';
 
 import { ApplicationEventBus } from './application/application.event-bus';
 import { ApplicationCommandFactory } from './application/application-command.factory';
